Implement modern versioning system with automated Git-tag based releases and dependency updates

+++ app/src/main/java/com/example/fitapp/services/WorkoutExecutionManager.kt
@@ -223,7 +223,7 @@ class WorkoutExecutionManager(
             actualRestTime = calculateActualRest(),
             formQuality = formQuality ?: 1.0f,
             perceivedExertion = rpe,
-            duration = actualSetDuration.toInt(),
+            duration = actualSetDuration,
             notes = notes
         )
